chore(refactor): refactor settings menu (previously profile menu)
