+++ locales/fr.json
@@ -79,7 +79,7 @@
   "NutFree": "Sans Noix",
   "LowCarb": "Faible en Glucides",
   "Household Size": "Taille du Foyer",
-  "Cuisine Variety Preference": "Préférence de Variété Culinaire",
+  "Cuisine Variety": "Variété Culinaire",
   "Repeat Often": "Répéter Souvent",
   "Balanced (Default)": "Équilibré (Par Défaut)",
   "Maximum Variety": "Variété Maximale",
@@ -406,6 +406,8 @@
   "Meal Planning": "Planification des Repas",
   "Recipe Management": "Gestion des Recettes",
   "Account & Billing": "Compte et Facturation",
+  "Billing": "Facturation",
+  "General": "Général",
   "Technical Support": "Support Technique",
   "TechnicalSupport": "Support Technique",
   "Accompaniment pairing": "Appariement des accompagnements",
@@ -751,6 +753,8 @@
   "Notifications": "Notifications",
   "Privacy & Security": "Confidentialité & Sécurité",
   "Logout": "Déconnexion",
+  "Log out": "Se déconnecter",
+  "Log out of all devices": "Se déconnecter de tous les appareils",
   "Account": "Compte",
   "Meals": "Repas",
   "Plan": "Plan",
