+++ src/routes/settings/mod.rs
@@ -1,5 +1,3 @@
 pub mod account;
-pub mod meal_preferences;
-// pub mod notifications;
-pub mod security;
-pub mod subscription;
+pub mod billing;
+pub mod general;
