--- src/routes/settings/security.rs
@@ -1,40 +1,0 @@
-// use axum::Form;
-use axum::extract::State;
-use axum::response::IntoResponse;
-// use serde::Deserialize;
-
-use crate::auth::AuthUser;
-use crate::routes::AppState;
-use crate::template::Template;
-use crate::template::filters;
-
-#[derive(askama::Template)]
-#[template(path = "settings-security.html")]
-pub struct SecurityTemplate {
-    // pub error_message: Option<String>,
-    pub current_path: String,
-    pub settings_path: String,
-    pub user: AuthUser,
-}
-
-pub async fn page(template: Template, user: AuthUser) -> impl IntoResponse {
-    template.render(SecurityTemplate {
-        // error_message: None,
-        current_path: "settings".to_owned(),
-        settings_path: "security".to_owned(),
-        user,
-    })
-}
-//
-// #[derive(Deserialize)]
-// pub struct ActionInput {
-//     pub email: String,
-// }
-
-pub async fn action(
-    _template: Template,
-    State(_app): State<AppState>,
-    // Form(input): Form<ActionInput>,
-) -> impl IntoResponse {
-    ""
-}
