+++ templates/kitchen.html
@@ -22,7 +22,7 @@
 <div ts-trigger="load" ts-req="/calendar/regenerate" ts-req-method="POST" ts-swap="skip"></div>
 {% endif %}
 {% if !user.is_premium() %}
-<div ts-trigger="load" ts-req="/settings/subscription/check" ts-req-method="POST" ts-swap="skip"></div>
+<div ts-trigger="load" ts-req="/settings/billing/check" ts-req-method="POST" ts-swap="skip"></div>
 {% endif %}
 
 <!-- Main Content -->
