+++ templates/partials/set-username-modal.html
@@ -1,6 +1,6 @@
 <div id="modal-set-username" class="fixed inset-0 bg-black/50 z-40 flex items-center justify-center p-4"
   ts-swap-push="body" ts-swap="append">
-  <form ts-req="/settings/account/set-username" ts-req-method="post" ts-target="parent #modal-set-username"
+  <form ts-req="/settings/general/set-username" ts-req-method="post" ts-target="parent #modal-set-username"
     class="bg-white rounded-lg shadow-xl max-w-md w-full">
     <!-- Body -->
     <div class="p-6">
@@ -20,8 +20,7 @@
         <div>
           <label class="block text-sm font-semibold text-gray-700 mb-2">{{ "Username"|t }}</label>
           <input autofocus autocomplete="off" type="text" name="username" minlength="3" maxlength="15" required
-            pattern="[a-zA-Z0-9_]+" title="{{ " Only alphanumeric characters allowed"|t }}" placeholder="{{ " Enter your
-            username, eg. john_doe"|t }}"
+pattern="[a-zA-Z0-9_]+" title="{{ "Only alphanumeric characters allowed"|t }}" placeholder="{{ "Enter your username, eg. john_doe"|t }}"
             class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-green-500 outline-none" />
           <p class="mt-2 text-sm text-gray-500">{{ "This will be visible to other users."|t }}</p>
         </div>
