+++ templates/partials/subscription-cancel-modal.html
@@ -60,7 +60,7 @@
     </ul>
 
     <div class="flex flex-col gap-3">
-      <button ts-req="/settings/subscription/cancel" ts-req-method="post" ts-swap="skip"
+      <button ts-req="/settings/billing/cancel" ts-req-method="post" ts-swap="skip"
         class="w-full px-4 py-2.5 bg-red-600 text-white font-semibold rounded-lg hover:bg-red-700 text-sm">
         Yes, Cancel Subscription
       </button>
