+++ templates/partials/subscription-payment.html
@@ -45,5 +45,5 @@
     </div>
   </div>
   {% endif %}
-  <button ts-req="/settings/subscription/update-payment" ts-target="body" ts-swap="append" class="text-green-600 hover:text-green-700 font-semibold text-sm">Update</button>
+  <button ts-req="/settings/billing/update-payment" ts-target="body" ts-swap="append" class="text-green-600 hover:text-green-700 font-semibold text-sm">Update</button>
 </div>
