+++ templates/settings-account.html
@@ -1,36 +1,34 @@
 {% extends "_settings.html" %}
-{% block title %}{{ "Profile & Settings"|t }} - imkitchen{% endblock %}
+{% block title %}{{ "Account"|t }} - imkitchen{% endblock %}
 
 {% block settings_content %}
 
-<!-- Account Information -->
 <div class="bg-white rounded-lg shadow-lg p-6">
-  <h2 class="text-xl font-bold mb-6">{{ "Account Information"|t }} <span class="px-3 py-1 bg-orange-500 text-white text-sm rounded-full font-semibold">🚀 Coming soon</span></h2>
+  <h2 class="text-xl font-bold mb-6">{{ "Account"|t }}</h2>
 
-  <form class="space-y-6">
-    <div>
-      <label class="block text-sm font-semibold text-gray-700 mb-2">{{ "Full Name"|t }}</label>
-      <input type="text" value="Sarah Mitchell"
-        class="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500" />
+  <div class="space-y-4 mb-6">
+    <div class="flex flex-col md:flex-row md:items-center justify-between gap-4">
+      <div class="flex-1">
+        <div class="text-sm text-gray-600">{{ "Log out of all devices"|t }}</div>
+      </div>
+      <a href="/logout"
+        class="w-full md:w-auto px-4 py-2 bg-gray-600 text-white font-semibold rounded-lg hover:bg-gray-700">
+        {{ "Log out"|t }}
+      </a>
     </div>
+  </div>
 
-    <div>
-      <label class="block text-sm font-semibold text-gray-700 mb-2">{{ "Email Address"|t }}</label>
-      <input type="email" value="sarah@example.com"
-        class="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500" />
-    </div>
-
-    <div>
-      <label class="block text-sm font-semibold text-gray-700 mb-2">{{ "Username"|t }}</label>
-      <input type="text" value="@sarah_cooks"
-        class="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500" />
-      <p class="text-sm text-gray-500 mt-1">{{ "Used for community recipe sharing"|t }}</p>
-    </div>
-
-    <button class="px-6 py-3 bg-green-600 text-white font-semibold rounded-lg hover:bg-green-700">
-      {{ "Save Changes"|t }}
-    </button>
-  </form>
+  <!-- <div class="space-y-4"> -->
+  <!--   <div class="flex flex-col md:flex-row md:items-center justify-between gap-4"> -->
+  <!--     <div class="flex-1"> -->
+  <!--       <div class="text-sm text-gray-600">{{ "Permanently delete your account and all data"|t }}</div> -->
+  <!--     </div> -->
+  <!--     <button onclick="if(confirm('{{ "Are you sure? This cannot be undone."|t }}')) alert('{{ "Account deletion initiated"|t }}')" -->
+  <!--       class="w-full md:w-auto px-4 py-2 bg-red-600 text-white font-semibold rounded-lg hover:bg-red-700"> -->
+  <!--       {{ "Delete Account"|t }} -->
+  <!--     </button> -->
+  <!--   </div> -->
+  <!-- </div> -->
 </div>
 
 {% endblock %}
