+++ templates/settings-billing.html
@@ -1,5 +1,5 @@
 {% extends "_settings.html" %}
-{% block title %}{{ "Profile & Settings"|t }} - imkitchen{% endblock %}
+{% block title %}{{ "Billing"|t }} - imkitchen{% endblock %}
 {% block page_scripts %}
 <script src="https://js.stripe.com/clover/stripe.js" defer></script>
 {% endblock %}
@@ -77,7 +77,7 @@
   </div>
   {% else %}
   <div class="mt-4 pt-4 border-t border-gray-200">
-    <button ts-req="/settings/subscription/cancel" ts-target="body" ts-swap="append"
+    <button ts-req="/settings/billing/cancel" ts-target="body" ts-swap="append"
       class="text-sm text-red-600 hover:text-red-700 font-semibold">
       Cancel Subscription
     </button>
@@ -88,7 +88,7 @@
 {% if subscription.is_active %}
 <div id="payment-method-panel" class="bg-white rounded-lg shadow-lg p-6">
   <h3 class="font-bold mb-4">Payment</h3>
-  <div ts-trigger="load" ts-req="/settings/subscription/payment-method" ts-req-method="post"
+  <div ts-trigger="load" ts-req="/settings/billing/payment-method" ts-req-method="post"
     class="flex items-center justify-between px-4 py-3">
     <div class="flex items-center gap-3">
       <div class="w-12 h-8 rounded-md shimmer"></div>
@@ -465,7 +465,7 @@
             elements: stripeElements,
             clientSecret: o.el.dataset.clientSecret,
             confirmParams: {
-              return_url: '{{ config.server.url }}/settings/subscription'
+              return_url: '{{ config.server.url }}/settings/billing'
             },
           }).then(({error}) => {
             setActionDisabled(false);
