+++ templates/settings-general.html
@@ -1,13 +1,13 @@
 {% extends "_settings.html" %}
-{% block title %}{{ "Profile & Settings"|t }} - imkitchen{% endblock %}
+{% block title %}{{ "General"|t }} - imkitchen{% endblock %}
 
 {% block settings_content %}
 
 <!-- Meal Preferences -->
 <div class="bg-white rounded-lg shadow-lg p-6">
-  <h2 class="text-xl font-bold mb-6">{{ "Meal Preferences"|t }}</h2>
+  <h2 class="text-xl font-bold mb-6">{{ "Preferences"|t }}</h2>
 
-  <form method="post" action="/settings/meal-preferences" ts-req="" ts-swap="skip" class="space-y-6" autocomplete="off">
+  <form method="post" action="/settings/general" ts-req="" ts-swap="skip" class="space-y-6" autocomplete="off">
     <div>
       <label class="block text-sm font-semibold text-gray-700 mb-3">{{ "Dietary Restrictions"|t }}</label>
       <div class="grid md:grid-cols-2 gap-3">
@@ -29,7 +29,7 @@
 
     <div>
       <label class="block text-sm font-semibold text-gray-700 mb-2">
-        {{ "Cuisine Variety Preference"|t }}
+        {{ "Cuisine Variety"|t }}
       </label>
       <input type="range" name="cuisine_variety_weight" min="0.1" max="1.0" step="0.1" value="{{ cuisine_variety_weight }}" class="w-full" />
       <div class="flex justify-between text-xs text-gray-500 mt-1">
