--- templates/settings-security.html
@@ -1,24 +1,0 @@
-{% extends "_settings.html" %}
-{% block title %}{{ "Profile & Settings"|t }} - imkitchen{% endblock %}
-
-{% block settings_content %}
-
-<!-- Danger Zone -->
-<div class="bg-white rounded-lg shadow-lg p-6 border-2 border-red-200">
-  <h2 class="text-xl font-bold mb-4 text-red-800">{{ "Danger Zone"|t }}</h2>
-
-  <div class="space-y-4">
-    <div class="flex flex-col md:flex-row md:items-center justify-between gap-4">
-      <div class="flex-1">
-        <div class="font-semibold">{{ "Delete Account"|t }}</div>
-        <div class="text-sm text-gray-600">{{ "Permanently delete your account and all data"|t }}</div>
-      </div>
-      <button onclick="if(confirm('{{ "Are you sure? This cannot be undone."|t }}')) alert('{{ "Account deletion initiated"|t }}')"
-        class="w-full md:w-auto px-4 py-2 bg-red-600 text-white font-semibold rounded-lg hover:bg-red-700">
-        {{ "Delete Account"|t }}
-      </button>
-    </div>
-  </div>
-</div>
-
-{% endblock %}
